backend/requirements.txt
Django==3.2.9
django-autoslug==1.9.8
django-filter==21.1
djangorestframework==3.12.4
djoser==2.1.0
drf-extra-fields==3.2.1
gunicorn==20.1.0
Pillow==8.4.0
python-dotenv==0.20.0
